Move constants file
